requirements.txt
uvicorn[standard]==0.16
fastapi==0.71
requests==2.27.1
python-decouple==3.5
